% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/Demuxmix-methods.R
\docType{class}
\name{Demuxmix-class}
\alias{Demuxmix-class}
\alias{Demuxmix}
\alias{show,Demuxmix-method}
\alias{summary,Demuxmix-method}
\alias{p.acpt}
\alias{p.acpt,Demuxmix-method}
\alias{p.acpt<-}
\alias{p.acpt<-,Demuxmix,numeric-method}
\title{A class representing a set of mixture models fitted to HTO data.}
\usage{
\S4method{show}{Demuxmix}(object)

\S4method{p.acpt}{Demuxmix}(object)

\S4method{p.acpt}{Demuxmix,numeric}(object) <- value

\S4method{summary}{Demuxmix}(object, ...)
}
\arguments{
\item{object}{An \code{Demuxmix} object.}

\item{value}{Value between 0 and 1 specifying the acceptance probability,
i.e., the minimum posterior probability required to assign a cell to
an HTO.}

\item{...}{Additional arguments forwarded to summary (ignored).}
}
\description{
An object of this class is returned by \code{\link{demuxmix}} and
stores a fitted mixture model for each hashtag of the given dataset and 
additional information about model fitting. Users should not directly
initialize this class but can apply various plotting functions and
extract information from an object of \code{Demuxmix}. For example,
\code{\link{dmmClassify}} returns the cell-to-HTO assignments obtained from
the models stored in the \code{Demuxmix} object. Please see the package's
vignette for how to work with an object of this class.
}
\details{
All matrices stored by \code{Demuxmix} have the same dimension and
  the same row and column names as the original matrix \code{hto} passed to
  \code{\link{demuxmix}}. The mixture models in slot \code{models} are
  stored in an internal class format.
}
\section{Functions}{
\itemize{
\item \code{show,Demuxmix-method}: Displays the object on the command line.

\item \code{p.acpt,Demuxmix-method}: Returns the acceptance probability \code{p.acpt}.

\item \code{p.acpt<-,Demuxmix,numeric-method}: Sets a new acceptance probability \code{p.acpt}.

\item \code{summary,Demuxmix-method}: Summarizes the classification results and
estimates error rates.
}}

\section{Slots}{

\describe{
\item{\code{models}}{A list of mixture models. One model per hashtag.}

\item{\code{outliers}}{A logical matrix of size HTO x cells identifying outlier 
values excluded from model fitting.}

\item{\code{clusterInit}}{A numeric matrix of size HTO x cells with the class
memberships used to initialize model fitting. A value of 1 corresponds to
the negative component and a value of 2 to the positive component.}

\item{\code{posteriorProb}}{A numeric matrix of size HTO x cells with the posterior
probabilities that a cell is positive for a hashtag.}

\item{\code{tailException}}{A logical matrix of size HTO x cells identifying
posterior probabilities that would be adjusted based on the exception
rules defined when calling \code{\link{demuxmix}} to correct inaccuracies
at the extreme tails of the mixture distributions. See
\code{\link{demuxmix}} for details.}

\item{\code{modelSelection}}{A \code{data.frame} with information about the model
selection process if parameter \code{model} was set to 'auto'. Empty 
\code{data.frame} if model was specified.}

\item{\code{parameters}}{A list with the \code{\link{demuxmix}} parameters
used to generate the model represented by this class.}
}}

\examples{
set.seed(2642)
simdata <- dmmSimulateHto(class=rbind(c(rep(TRUE, 220), rep(FALSE, 200)),
                                      c(rep(FALSE, 200), rep(TRUE, 220))))

dmm <- demuxmix(simdata$hto, rna=simdata$rna, p.acpt=0.9)
p.acpt(dmm)
head(dmmClassify(dmm))
dmm@parameters

}
\seealso{
\code{\link{dmmClassify}} to obtain classification results.
  \code{\link{plotDmmHistogram}}, \code{\link{plotDmmScatter}},
  \code{\link{plotDmmPosteriorP}}, and \code{\link{dmmOverlap}} to assess
  the model fit.
}
