% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demuxmix-package.R
\docType{package}
\name{demuxmix-package}
\alias{demuxmix}
\alias{demuxmix-package}
\title{demuxmix: Demultiplexing oligo-barcoded scRNA-seq data using regression mixture models}
\description{
A package for demultiplexing single-cell sequencing experiments of pooled cells labeled with barcode oligonucleotides. The package implements methods to fit regression mixture models for a probabilistic classification of cells, including multiplet detection. Demultiplexing error rates can be estimated, and methods for quality control are provided.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/huklein/demuxmix}
  \item Report bugs at \url{https://github.com/huklein/demuxmix/issues}
}

}
\author{
\strong{Maintainer}: Hans-Ulrich Klein \email{hansulrich.klein@gmail.com} (\href{https://orcid.org/0000-0002-6382-9428}{ORCID})

}
\keyword{internal}
