% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{dmmSummary}
\alias{dmmSummary}
\alias{dmmSummary,data.frame-method}
\title{Summarize classification results from demuxmix}
\usage{
dmmSummary(dmmResults, p.acpt)
}
\arguments{
\item{dmmResults}{A "results" \code{data.frame} generated by
\code{\link{demuxmix}}.}

\item{p.acpt}{An optional value between 0 and 1 defining the acceptance
probability that must be reached in order to assign a cell to a hashtag.
If missing, the original acceptance probability passed to
\code{\link{demuxmix}} when generating \code{dmmResults} is used.}
}
\value{
A \code{data.frame} with one row per class showing the number
  of cells in the class (NumObs), the relative frequency of the class
  (RelFreq), the median probability that a cell assigned to the class
  is actually a member of the class (MedProb), the estimated number of
  cells falsely assigned to the class (ExpFPs), and the corresponding
  false discovery rate (FDR).
}
\description{
This method takes the demultiplexing results from an HTO experiment
returned by \code{\link{demuxmix}} and returns a \code{data.frame}
summarizing the classification results and classification certainty.
}
\details{
Results are summarized for the individual hashtags, for all
  singletons combined, for all multiplets, and for the negative class.
  Relative frequencies are calculated after excluding the "uncertain"
  class. The estimated number of false positive cells and the
  estimated FDR are based on several assumptions, one of which is the
  independence of the HTO counts from the different hashtags. This
  assumption unlikely for real data where HTO counts of different hashtags
  are obtained from the same cells. Error rates for singletons can be 
  expected to be conservative whereas error rates for multiplets and
  negative cells are likely underestimated.
}
\examples{
set.seed(2642)
simdata <- dmmSimulateHto(class=rbind(c(rep(TRUE, 220), rep(FALSE, 200)),
                                      c(rep(FALSE, 200), rep(TRUE, 220))))

dmm <- demuxmix(simdata$hto, rna=simdata$rna, p.acpt=0.9)
dmmSummary(dmm$results)
dmmSummary(dmm$results, p.acpt=0.5)

}
\seealso{
\code{\link{demuxmix}}
}
