% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{demuxmix}
\alias{demuxmix}
\alias{demuxmix,matrix,missing-method}
\alias{demuxmix,matrix,numeric-method}
\alias{demuxmix,Matrix,missing-method}
\alias{demuxmix,Matrix,numeric-method}
\title{Demultiplexing using mixture models}
\usage{
demuxmix(
  hto,
  rna,
  p.acpt = 0.9^nrow(hto),
  model = "auto",
  alpha = 0.9,
  beta = 0.9,
  correctTails = TRUE,
  tol = 10^-5,
  maxIter = 100,
  k.hto = 1.5,
  k.rna = 1.5
)
}
\arguments{
\item{hto}{A matrix of HTO counts where each row corresponds to a hashtag
and each column to a cell barcode. The matrix must have unique row names.}

\item{rna}{An optional numeric vector with the number of genes detected in
the RNA library for each cell. Same length as columns in \code{object}.
If missing, parameter \code{model} must be set to "naive".}

\item{p.acpt}{Acceptance probability that must be reached in order to
assign a cell to a hashtag. Cells with lower probabilities are classified
as "uncertain".}

\item{model}{A character specifying the type of mixture model to be used.
Either "naive", "regpos", "reg" or "auto". The last three options
require parameter \code{rna} to be specified. "auto" selects the best
model based on maximum difference between the two components of the
mixture models.}

\item{alpha}{Threshold defining the left tail of the mixture
distribution where cells should not be classified as "positive". Threshold
must be between 0 and 1. See details.}

\item{beta}{Threshold for defining the right tail of the mixture
distribution where cells should not be classified as "negative". Threshold
must be between 0 and 1. See details.}

\item{correctTails}{If \code{TRUE}, cells meeting the threshold defined by
\code{alpha} (\code{beta}) are classified as "negative" ("positive") even
if the mixture model suggests a different classification. See details.}

\item{tol}{Convergence criterion for the EM algorithm fitting the mixture
model(s). The algorithm stops when the relative increase of the log
likelihood is less than or equal to \code{tol}.}

\item{maxIter}{Maximum number of iterations for the EM algorithm and
for the alternating iteration process fitting the NB regression models
within each EM iteration.}

\item{k.hto}{Factor to define outliers in the HTO counts. Among cells
positive for the hashtag based on initial clustering, HTO counts
larger than the 0.75 quantile + \code{k.hto} * IQR are considered
outliers. See details.}

\item{k.rna}{Factor to define outliers in the numbers of detected genes.
Numbers of detected genes larger than the 0.75 quantile +
\code{k.rna} * IQR are considered outliers. See details.}
}
\value{
\code{demuxmix} returns an object of class \code{\link{Demuxmix}}.
  Classification results can be extracted with \code{\link{dmmClassify}}.
  Various plot methods (see below) are available to assess the model fit.
}
\description{
This methods uses mixture models as a probabilistic framework to assign
cells to hashtags and to identify multiplets based on counts obtained from
a hastag oligo (HTO) library. If the numbers of detected genes
from the corresponding RNA library are passed as second argument,
regression mixture models can be used, which usually improves the
classification accuracy by leveraging the relationship between HTO and RNA
counts.
}
\details{
The single cell dataset should undergo basic filtering to
  remove low quality or empty droplets before calling this function,
  but the HTO counts themselves should not be transformed or pre-processed
  otherwise. The number of detected genes passed via the optional argument
  \code{rna} is typically defined as the number of genes in the RNA library
  with at least one read.
  
  The method fits a two-component negative binomial mixture model for each
  hashtag. The type of mixture model used can be specified by \code{model}.
  "naive" fits a standard mixture model. "reg" fits a regression mixture
  model using the given number of detected genes (\code{rna}) as covariate
  in the regression model. "regpos" uses a regression model only for the
  positive but not for the negative component. If model is set to "auto",
  all three models are fitted and the model with the largest difference
  between the two components (best separation) is selected. In most HTO
  datasets, regression mixture models perform better.
  
  The \code{demuxmix} method consists of 3 steps, which can be tuned
  by the respective parameters. The default settings work well for a wide
  range of datasets and should only be changed if any issues arise during
  model fitting and quality control. An exception is the acceptance
  probability \code{p.acpt}, which may be set to smaller or larger value
  depending on the desired tradeoff between number of unclassified/discarded
  cells and error rate. Steps 1 and 2 are executed for each hashtag
  separately using only the hto counts for that hashtag; step 3 classifies
  the cells based on the results from all hashtags. Therefore, parameters
  affecting steps 1 and 2 (incl. \code{model}) can be specified for each
  hashtag using a vector with one element per hashtag. Shorter vectors will
  be extended.
  
  \enumerate{
    \item Preprocessing (\code{k.hto}, \code{k.rna}). Cells are clustered
    into a negative and a positive group based on the hto counts of the
    hashtag using k-means. Cells in the positive group with hto counts
    larger than \code{k.hto} times the IQR of the hto counts in the positive
    group are marked as outliers. Outliers are still classified but will
    not be used to fit the mixture model for this hashtag in step 2. If
    the parameter \code{rna} is given and the \code{model} is "reg" or
    "regpos", all cells (both groups) with number of detected genes larger
    than k.rna times the IQR are marked as outliers, too, since these cells 
    could affect the fitting of the regression model negatively. If more
    than 5\% of the cells are marked as outliers, a warning message is
    printed and larger values for \code{k.hto} and \code{k.rna}
    might be preferable. If the model fit seems to be affected by a few
    large values (very high variance of the positive component), smaller
    values should be chosen.
    
    \item Model fitting (\code{model}, \code{alpha}, \code{beta},
    \code{correctTails}, \code{tol}, \code{maxIter}). An EM algorithm is
    used to fit a mixture model to the HTO counts which were not marked as
    outliers in step 1. \code{maxIter} defines the maximum number of
    iterations of the EM algorithm, and, if \code{model} is "reg", "regpos"
    or "auto", it also defines the maximum number of iterations to fit the
    negative binomial regression models within each EM iteration. \code{tol}
    defines the convergence criterion for the EM algorithm. The algorithm
    stops if \eqn{\Delta LL/LL \le} \code{tol}. After the mixture model has
    been fitted, the posterior probability that cell X is positive for
    the hashtag \eqn{P(X = pos)} is calculated. Depending on the given data,
    these probabilities can be inaccurate at the far tails of the mixture
    distribution. Specifically, a positive component with large variance can
    have a larger value close to zero than the negative component, if the
    negative component is narrow and shifted to the right due to ambient
    hashtags. If \code{correctTails} is \code{TRUE}, the following two rules
    are applied to avoid false classifications at the at the far tails.
    First, if a cell X is classified as positive based on the posterior
    probability, but the probability to detected more than the observed
    \eqn{h_x} hashtags in a negative cell is \eqn{P(H \ge h_x | neg)} >
    \code{alpha}, then \eqn{P(X = pos)} is set to 0 (left tail). Second,
    if a cell X is classified as negative, but \eqn{P(H \le h_x | pos)} >
    \code{beta}, \eqn{P(X = pos)} is set to 1 (right tail). For most
    datasets, these rules will not apply and it's recommended not to change
    these values. If \code{correctTails} is \code{FALSE}, posterior
    probabilities will not be altered, but potential problems at the tails
    will still be logged in the slot \code{tailException} of the returned
    object.
    
    \item Classification (\code{p.acpt}). The posterior probabilities
    obtained from the models fitted to each hashtag separately are 
    used to calculate the most likely class for each cell. The following
    classes are considered: one class for each hashtag (singletons), one
    class for each possible multiplet, and a negative class representing
    droplets negative for all hashtags (i.e. empty droplets or droplets
    containing only cell debris). Each cell is assigned to the most
    likely class unless the probability is smaller than \code{p.acpt},
    in which case the cell is assigned to the class "uncertain". Apply the
    method \code{\link{dmmClassify}} to an object returned by \code{demuxmix}
    to obtain the classification results. The acceptance probability
    can be changed after running \code{demuxmix} using
    \code{\link{p.acpt<-}}.
  }
}
\examples{
set.seed(2642)
simdata <- dmmSimulateHto(class=rbind(c(rep(TRUE, 220), rep(FALSE, 200)),
                                      c(rep(FALSE, 200), rep(TRUE, 220))))

dmm <- demuxmix(simdata$hto, model="naive")
dmm
table(dmmClassify(dmm)$HTO, simdata$groundTruth)

dmmreg <- demuxmix(simdata$hto, rna=simdata$rna)
dmm
table(dmmClassify(dmmreg)$HTO, simdata$groundTruth)
summary(dmmreg)

p.acpt(dmmreg) <- 0.5
summary(dmmreg)

dmmOverlap(dmmreg)
\donttest{
plotDmmHistogram(dmmreg)
plotDmmScatter(dmmreg, hto="HTO_1")}

}
\seealso{
\code{\link{dmmClassify}} to extract the classification results
  and \code{\link{summary}} to summarize the results.
  \code{\link{plotDmmHistogram}}, \code{\link{plotDmmScatter}},
  \code{\link{plotDmmPosteriorP}}, and \code{\link{dmmOverlap}} to assess
  the model fit.
}
