% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{dmmClassify}
\alias{dmmClassify}
\alias{dmmClassify,Demuxmix-method}
\title{Return classification results from a Demuxmix object}
\usage{
dmmClassify(object)
}
\arguments{
\item{object}{An object of class \code{\link{Demuxmix}}.}
}
\value{
A \code{data.frame} with 3 columns and one row for each cell
  in the dataset. The first column gives the class (HTO) the cell has been
  assigned to. The second column contains the posterior probability. And the
  third column specifies the type of the assigned class, i.e., singlet,
  multiplet, negative or uncertain.
}
\description{
This method uses the posterior probabilities from the given demuxmix model
to assign each cell to the most likely class, either a single HTO,
a combination of HTOs (multiplet) or the negative class (not hashtagged
cells, empty droplets, cell debris). If the assignment cannot be made with
certainty above a defined threshold, the cell is labeled as "uncertain".
}
\details{
A cell is labeled as "uncertain" if the posterior probability of
  the most likely class is smaller than the threshold \code{p.acpt}, which
  is stored in the given \code{\link{Demuxmix}} object. The acceptance
  probability \code{p.acpt} can be inspected and set to a different value
  by applying the getter/setter method \code{\link{p.acpt}} to the
  \code{\link{Demuxmix}} object before calling this method. The method
  \code{\link{summary}} is useful to investigate the effect of and to
  estimate error rates for different values of \code{p.acpt}.
}
\examples{
set.seed(2642)
simdata <- dmmSimulateHto(class=rbind(c(rep(TRUE, 220), rep(FALSE, 200)),
                                      c(rep(FALSE, 200), rep(TRUE, 220))))

dmm <- demuxmix(simdata$hto, rna=simdata$rna)
head(dmmClassify(dmm))
table(dmmClassify(dmm)$HTO, simdata$groundTruth)

p.acpt(dmm) <- 0.5
sum(dmmClassify(dmm)$HTO == "uncertain")
p.acpt(dmm) <- 0.9999
sum(dmmClassify(dmm)$HTO == "uncertain")

}
\seealso{
\code{\link{demuxmix}}
}
