% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{dmmOverlap}
\alias{dmmOverlap}
\alias{dmmOverlap,list-method}
\title{Calculate area intersected by two components of a mixture model}
\usage{
dmmOverlap(object, hto, tol = 0.001)
}
\arguments{
\item{object}{A mixture model or a list of mixture models as
returned by \code{\link{demuxmix}}.}

\item{tol}{The maximum acceptable error when calculating the
area.}
}
\value{
A numeric vector with the intersection area for each given
  mixture model.
}
\description{
\code{dmmOverlap} integrates over the area intersected by the two
components of the given mixture model. The integral should be
close to 0 if the HTO labeling experiment was successful.
}
\details{
The area under both the negative and positive component is an
  informative quality metric for the hashtag labeling efficiency. Values
  under 0.03 can be considered as good, values larger than 0.1 are
  problematic.
  
  The definition of the area is not obvious for a regression mixture
  model since the distributions' means depend on the covariate, i.e.,
  the number of detected genes in the RNA library. This method 
  calculates the weighted mean number of detected genes in the cells
  for each component, which are then used to calculate the expected
  number of HTO counts for the negative and positive component.
}
\examples{
set.seed(2642)
simdata <- dmmSimulateHto(class=rbind(c(rep(TRUE, 220), rep(FALSE, 200)),
                                      c(rep(FALSE, 200), rep(TRUE, 220))))

dmm <- demuxmix(simdata$hto, p.acpt=0.9)
dmmOverlap(dmm$model)

dmmreg <- demuxmix(simdata$hto, rna=simdata$rna, p.acpt=0.9)
dmmOverlap(dmmreg$model)
dmmOverlap(dmmreg$model[["HTO_1"]])

}
\seealso{
\code{\link{demuxmix}}
}
