% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plotDmmHistogram}
\alias{plotDmmHistogram}
\alias{plotDmmHistogram,list-method}
\title{Plotting a histogram with mixture densities}
\usage{
plotDmmHistogram(model, quantile = 0.95, binwidth = 5)
}
\arguments{
\item{model}{A mixture model or a list of mixture models returned by
\code{\link{demuxmix}}.}

\item{quantile}{Quantile of the mixture distribution which is used as
right limit of the x axis of the plot.}

\item{binwidth}{Width of the bins of the histogram.}
}
\value{
An object of class \code{ggplot} is returned. Or, if multiple
  mixture models are passed to this function, an object of class
  \code{gtable} is returned.
}
\description{
This methods plots the mixture density and the components' densities
on top of a histogram of the HTO counts used to fit the mixture model.
The mixture model must be generated by \code{\link{demuxmix}}.
}
\details{
A histogram overlaid with the densities from the mixture model
  is often helpful to assess the mixture model fit. How such a plot
  should generated for regression mixture models is not obvious. This
  method calculates the weighted mean number of detected genes in the
  cells for each component, which are then used to calculate the
  expected number of HTO counts for the negative and positive component.
  The HTO counts shown in the histogram are adjusted to account for
  different numbers of detected genes by replacing the original HTO
  counts with the expected counts given the mean number of detected
  genes plus the residual counts from the regression model.
  
  Sometimes it is useful to zoom into the plot to obtain a better view
  of the fit. To restrict the plot to a certain range on the x or y axis,
  the method \code{\link[ggplot2]{coord_cartesian}} from the \code{ggplot2}
  package should be used (see examples).
}
\examples{
set.seed(2642)
simdata <- dmmSimulateHto(class=rbind(c(rep(TRUE, 220), rep(FALSE, 200)),
                                      c(rep(FALSE, 200), rep(TRUE, 220))))

dmm <- demuxmix(simdata$hto, p.acpt=0.9)
\donttest{plotDmmHistogram(dmm$model)}

dmmreg <- demuxmix(simdata$hto, rna=simdata$rna, p.acpt=0.9)
\donttest{plotDmmHistogram(dmmreg$model)}

p <- plotDmmHistogram(dmmreg$model[1])
\donttest{p + ggplot2::coord_cartesian(xlim=c(25, 100), ylim=c(0, 0.01))}

}
\seealso{
\code{\link{demuxmix}}
}
