% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plotDmmPosteriorP}
\alias{plotDmmPosteriorP}
\alias{plotDmmPosteriorP,list-method}
\title{Plotting a histogram of posterior probabilities}
\usage{
plotDmmPosteriorP(model, log = FALSE, bins = 50)
}
\arguments{
\item{model}{A mixture model or a list of mixture models returned
\code{\link{demuxmix}}.}

\item{log}{Logical value indicating whether a logarithmic scale
should be used for the frequencies (y axis).}

\item{bins}{The number of bins of the histogram.}
}
\value{
An object of class \code{ggplot} is returned. Or, if multiple
  mixture models are passed to this function, an object of class
  \code{gtable} is returned.
}
\description{
This methods plots a histogram of posterior probabilities obtained
from the given mixture model. The probabilities indicate whether the
cell originates from the positive component of the mixture model (i.e.
tagged with the hashtag) or from the negative component. The mixture
model passed to this function must be generated by
\code{\link{demuxmix}}.
}
\details{
The histogram visualizes how well the positive cells can be
  separated from the cells negative for the hashtag. Ideally, the
  histogram shows many cells with a posterior probability very close
  to 0 and many cells close to 1, but no or very few cells with
  probabilities somewhere in between. The histogram can be useful to
  guide the choice of the acceptance probability \code{p.acpt} passed
  to \code{\link{demuxmix}}.
}
\examples{
set.seed(2642)
simdata <- dmmSimulateHto(class=rbind(c(rep(TRUE, 220), rep(FALSE, 200)),
                                      c(rep(FALSE, 200), rep(TRUE, 220))))

dmm <- demuxmix(simdata$hto, p.acpt=0.9)
\donttest{plotDmmPosteriorP(dmm$model)}

dmmreg <- demuxmix(simdata$hto, rna=simdata$rna, p.acpt=0.9)
\donttest{plotDmmPosteriorP(dmmreg$model)}
\donttest{plotDmmPosteriorP(dmmreg$model[1])}

}
\seealso{
\code{\link{demuxmix}}
}
