% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plotDmmScatter}
\alias{plotDmmScatter}
\alias{plotDmmScatter,list-method}
\title{Plotting RNA features versus HTO counts}
\usage{
plotDmmScatter(model, log = TRUE, pointsize = 1.2)
}
\arguments{
\item{model}{A regression mixture model or a list of regression mixture
models returned by \code{\link{demuxmix}}.}

\item{log}{Logical value indicating whether both HTO counts and number
of RNA features should be log transformed.}

\item{pointsize}{Numeric value specifying the size of the points.}
}
\value{
An object of class \code{ggplot} is returned. Or, if multiple
  mixture models are passed to this function, an object of class
  \code{gtable} is returned.
}
\description{
This methods plots the numbers of detected RNA features versus the numbers
of HTO counts and indicates the posterior probability obtained from the
mixture model by different colors. The mixture model passed to this
function must be a regression mixture model generated by
\code{\link{demuxmix}}.
}
\details{
The scatterplot produced by this method is helpful to assess
  the relation between the number of RNA features detected and the number
  of HTO counts obtained for a cell. A positive association is usually
  visible for the positive cells with a large posterior probability
  indicated by red color.
}
\examples{
set.seed(2642)
simdata <- dmmSimulateHto(class=rbind(c(rep(TRUE, 220), rep(FALSE, 200)),
                                      c(rep(FALSE, 200), rep(TRUE, 220))))

dmmreg <- demuxmix(simdata$hto, rna=simdata$rna, p.acpt=0.9)
\donttest{plotDmmScatter(dmmreg$model)}
\donttest{plotDmmScatter(dmmreg$model[1])}

}
\seealso{
\code{\link{demuxmix}}
}
